+++ SPTK/sptk.c
@@ -86,8 +86,8 @@ DP_SETTINGS *initDpSettings()
     dp->recordingFrameAfter = 1;
     dp->recordingFrameBefore = 5;
     dp->recordingMaxFiles = 10;
-    dp->recordingType = 0;
-    dp->mark_delimeter = 10.0;
+    dp->recordingType = 3;
+    dp->mark_delimeter = 5.0;
     dp->mark_level = 50.0;
     dp->mark_labels = "F (Failure), D (Insufficient), C (Mediocre), B (Satisfactory), A (Excellent)";
     dp->smooth_frame = 1;