+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2018-02-10
+
+### Changed
+- Settings window move Similarity Measure Calculation to main settings
+
+### Removed
+- tab 'more' from graph window
+- Remove unused settings
+
+## [0.2.2] - 2018-02-09
+
+### Changed
+- Settings window changed titles and labels
+- Graph window changed titles and labels
+
 ## [0.2.1] - 2018-02-04
 
 ### Added